backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
dicebear==1.2.4
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==16.4.0
idna==3.4
Pillow==9.4.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
